% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_htna.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap generic}
\usage{
bootstrap(x, ...)
}
\arguments{
\item{x}{An object to bootstrap.}

\item{...}{Arguments forwarded to the method.}
}
\value{
An object whose structure is method-defined.
}
\description{
S3 generic dispatched on the class of \code{x}. Provided so \code{bootstrap(net)}
works directly on an htna network (see \code{\link[=bootstrap_htna]{bootstrap_htna()}}).
}
