% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_htna.R
\docType{data}
\name{htna_palette}
\alias{htna_palette}
\title{Default HTNA colour palette}
\format{
An object of class \code{character} of length 6.
}
\usage{
htna_palette
}
\value{
A character vector of 6 hex colour codes.
}
\description{
A colour palette for up to 6 actor groups, used by \code{\link[=plot_htna]{plot_htna()}} when
no explicit colours are supplied.
}
\examples{
htna:::htna_palette
}
\keyword{internal}
