% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_plot_htna.R
\name{plot_sequences}
\alias{plot_sequences}
\title{Sequence plot generic}
\usage{
plot_sequences(x, ...)
}
\arguments{
\item{x}{An object to plot.}

\item{...}{Forwarded to the method.}
}
\value{
Method-defined.
}
\description{
S3 generic dispatched on \code{x}. Calling \code{plot_sequences(net)} on an htna
network forwards to \code{\link[=sequence_plot_htna]{sequence_plot_htna()}}.
}
