% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stability.R
\name{plot.htna_stability_group}
\alias{plot.htna_stability_group}
\title{Plot Method for htna Stability Groups}
\usage{
\method{plot}{htna_stability_group}(x, ...)
}
\arguments{
\item{x}{An object of class \code{htna_stability_group} from \code{\link[=centrality_stability_htna]{centrality_stability_htna()}}.}

\item{...}{Additional arguments passed to the plotting method.}
}
\value{
A combined plot or list of plots for each group.
}
\description{
S3 method for plotting grouped htna stability results.
}
